requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
nltk==3.9.4
numpy==2.4.4
pandas==3.0.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
regex==2026.4.4
six==1.17.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
Werkzeug==3.1.8
